% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_envir_setup.r
\name{Parallel_envir_setup}
\alias{Parallel_envir_setup}
\title{Set Up or Stop Parallel Environment}
\usage{
Parallel_envir_setup(
  spare_korez = 4,
  action = "setup",
  verbose = FALSE,
  clean = FALSE
)
}
\arguments{
\item{spare_korez}{Numeric. The number of cores to reserve for other tasks.
The default value is 4. The number of cores used for
parallel computation will be the total number of cores
minus this value.}

\item{action}{Character. Specifies the action to be taken. Use "setup" to
initialize the environment or "stop" to stop the cluster.
The default value is "setup".}

\item{verbose}{Logical. If TRUE, prints debug information about the number
of workers registered. The default is FALSE.}

\item{clean}{Logical. If TRUE, prevents loading of packages into the global
environment. The default is FALSE.}
}
\description{
This function configures or stops a parallel computing environment based on
the specified action. It sets up a cluster with a specified number of cores,
or stops the cluster if the action is "stop". The function can also load
required packages into the main environment based on the \code{clean} parameter.
}
\details{
When \code{action} is "setup", the function detects the total number of cores
available on the machine, subtracts the number of cores specified by
\code{spare_korez}, and creates a parallel cluster with the remaining cores. It
then registers this cluster for use with \verb{\%dopar\%} from the \code{foreach} package.
When \code{action} is "stop", the function stops the previously created cluster.
}
\examples{
# Set up the parallel environment with 4 cores reserved
Parallel_envir_setup(spare_korez = 4)

# Stop the parallel environment
Parallel_envir_setup(action = "stop")

}
